+++ components/analytics/shake2report/impl/src/main/java/com/flipperdevices/analytics/shake2report/impl/composable/states/ComposableReportPending.kt
@@ -21,8 +21,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.analytics.shake2report.impl.R
 import com.flipperdevices.analytics.shake2report.impl.composable.components.ComposableCheckBox
-import com.flipperdevices.analytics.shake2report.impl.composable.components.ComposableReportTextField
 import com.flipperdevices.core.ui.ktx.elements.ComposableFlipperButton
+import com.flipperdevices.core.ui.ktx.elements.ComposableReportTextField
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
@@ -78,7 +78,7 @@ private fun ComposableReportPendingFields(
     verticalArrangement = Arrangement.spacedBy(16.dp)
 ) {
     val placeHolderStyle = LocalTypography.current.bodyR14.copy(
-        color = LocalPallet.current.borderViewReportBug
+        color = LocalPallet.current.reportBorder
     )
     val titleStyle = LocalTypography.current.buttonB16.copy(
         color = LocalPallet.current.text100