+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/model/FapMetaInformation.kt
@@ -4,5 +4,5 @@ import com.flipperdevices.core.data.SemVer
 
 data class FapMetaInformation(
     val version: SemVer,
-    val sizeBytes: Long
+    val sizeBytes: Long?
 )